Initial visually working Transform architecture
This commit makes a single World entity responsible for the scene, and
makes that world entity the parent of all scene objects. Each tile in a
tilemap becomes a child entity with a controller that manages that tiles
drawable. For performance reasons, this may be changed later to have a
"Tilemap" controller, so that we don't need to be managing 900 entities
JUST for the background world.

Also, big changes to how controllers are handled, as the previous
unstructured method was causing massive memory issues. Changing it to a
class had a HUGE positive impact, dropping the memory consumption by
85% and ending the out-of-memory issues.