Resolve worker pid when worker server starts
Delays resolving the createRoom request until the room worker has successfully
been created and begun the websocket server. This prevents the player getting
redirected to a room before the server is actually started. There mightly be a
better way to handle this, for example delaying rendering the play route until
the server is listening,, but this works for now.